+++ examples/worker_producer/common.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 use Ds\Map;
 use Zlikavac32\BeanstalkdLib\Client\ProtocolClient;
 use Zlikavac32\BeanstalkdLib\Client\TubeConfiguration\StaticTubeConfiguration;
-use Zlikavac32\BeanstalkdLib\Client\TubeConfiguration\TubeMapConfigurationFactory;
 use Zlikavac32\BeanstalkdLib\Serializer;
 
 require_once __DIR__.'/../common.php';
@@ -71,8 +70,6 @@ const TUBE_INDEX_PROJECT_COMMIT = 'index.project_commit';
 
 $serializer = new ProjectCommitSerializer();
 
-$tubeConfigurationFactory = new TubeMapConfigurationFactory(new Map([
+$client = new ProtocolClient($protocol, new Map([
     TUBE_INDEX_PROJECT_COMMIT => new StaticTubeConfiguration(0, 1024, 300, 3600, $serializer),
 ]));
-
-$client = new ProtocolClient($protocol, $tubeConfigurationFactory);