+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* **[REMOVED]** `Zlikavac32\BeanstalkdLib\Client\TubeConfiguration\TubeConfigurationFactory`
 * **[CHANGED]** `Zlikavac32\BeanstalkdLib\JobDispatcher\TubeMapJobDispatcher` does not ignore `default` tube if there is no explicitly watched tube
 * **[ADDED]** `Zlikavac32\BeanstalkdLib\Protocol\TraceableProtocol` and `Zlikavac32\BeanstalkdLib\Adapter\PHPUnit\Constraint\OrderedTracesExistInProtocol` to help with testing